+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add option for `logs get` to download logs in parallel using the `--threads X` option.
 * Symlink `attach shell` command to `cluster attach` command.
 * Add option in `tls generate` to have different bit sizes.
+* If `-f` force option is not specified in the `destroy` commands, ask for confirmation via `stdin`.
 
 #### 6.0.3
 * Handle bug in IP discovery on `docker` backend with custom networks.